fix + modelmapper

+++ collection and processing of sensor data/1dz-server-template/src/main/java/hr/fer/tel/rassus/server/beans/DTO/ReadingResponseDTO.java
@@ -13,17 +13,17 @@ public class ReadingResponseDTO {
 
     private Long id;
 
-    private float temperature;
+    private double temperature;
 
-    private float pressure;
+    private double pressure;
 
-    private float humidity;
+    private double humidity;
 
-    private float co;
+    private double co;
 
-    private float no2;
+    private double no2;
 
-    private float so2;
+    private double so2;
 
     private Long sensorId;
 }
