+++ collection and processing of sensor data/1dz-server-template/src/main/java/hr/fer/tel/rassus/server/beans/DTO/SensorDTO.java
@@ -11,12 +11,12 @@ import lombok.Setter;
 @Setter
 public class SensorDTO {
 
-    private float latitude;
+    private double latitude;
 
-    private float longitude;
+    private double longitude;
 
     private String ip;
 
-    private long port;
+    private double port;
 
 }
