+++ collection and processing of sensor data/1dz-server-template/src/main/java/hr/fer/tel/rassus/server/beans/DTO/SensorResponseDTO.java
@@ -15,13 +15,13 @@ public class SensorResponseDTO {
     @NotNull
     private long id;
 
-    private float latitude;
+    private double latitude;
 
-    private float longitude;
+    private double longitude;
 
     private String ip;
 
-    private long port;
+    private double port;
 
 }
