+++ collection and processing of sensor data/1dz-server-template/src/main/java/hr/fer/tel/rassus/server/beans/Sensor.java
@@ -20,9 +20,9 @@ public class Sensor {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private float latitude;
+    private double latitude;
 
-    private float longitude;
+    private double longitude;
 
     private String ip;
